requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
PyMySQL
python-dotenv
Werkzeug
email-validator